requirements.txt
chardet==5.2.0
contourpy==1.3.1
cycler==0.12.1
fonttools==4.55.8
kiwisolver==1.4.8
matplotlib==3.10.0
mysql-connector-python==8.3.0
numpy==2.2.2
packaging==24.2
pillow==11.1.0
pyparsing==3.2.1
PyQt6==6.8.0
PyQt6-Charts==6.8.0
PyQt6-Charts-Qt6==6.8.1
PyQt6-Qt6==6.8.1
PyQt6_sip==13.10.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
reportlab==4.3.0
six==1.17.0
